LayoutTests/batterystatus/updates-expected.txt

Tests that updates to the battery event causes new events to fire.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS battery.charging is false
PASS battery.chargingTime is Infinity
PASS battery.dischargingTime is 6000
PASS battery.level is 0.7
PASS battery.charging is true
PASS battery.chargingTime is 7000
PASS battery.dischargingTime is Infinity
PASS battery.level is 0.3
PASS successfullyParsed is true

TEST COMPLETE
